Move html content plugin implementation into the Robotskirt plugin
We use a lot of markdown-type functionality in the html implementation. Eg.
Maybe we'll want to keep codeblocks and highlighting. So, instead of duplicating
the code across 2 plugins, I just integrated it as a sub-class of
RobotskirtPage.

The only difference between HtmlPage and RobotskirtPage is that HtmlPage does
not perform markdown rendering. It parses the yaml, renders the template, and
passes the html body directly to the rendered output.